refactor: separate camera and renderer
